gfs2: eliminate target parameter from finish_xmote
Function finish_xmote took a parameter target, which was often set as
from gl->gl_req. In glock_work_func it used the value of gl_reply, the
reply from the dlm response to the request. This patch simplifies function
finish_xmote by eliminating the target parameter in favor of just using
gl->gl_req. This simplifies finish_xmote and prepares for it to ecome a
state in a new state machine in future patches.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>