gfs2: eliminate gh parameter from do_xmote
Before this patch, function do_xmote was passed in a holder record that
was determined by find_first_waiter. There were a few exceptions in the
case of function run_queue. These exceptions were moved to do_xmote and
the parameter eliminated. This simplified do_xmote and makes way for
more improvements to the glock state machine.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>