gfs2: make GLF_BLOCKING flag managed entirely in lock_dlm
The GLF_BLOCKING flag is used for debug purposes (lock flag in traces and
lock dumps) and keeping glock statistics. Instead of making do_xmote manage
the flag, it is pushed into lock_dlm as it makes the requests.

This patch is a baby step toward making the glock state machine better by
reducing the need to access holder records in function do_xmote.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>