gfs2: move much of glock_work_func to state_run_queue
Before this patch, glock_work_func did a bunch of work related to
requeuing the work when needed, being careful to drop the required
additional glock references. In the middle, it called the state
machine to run the run_queue state.

This patch moves a bunch of that functionality from glock_work_func
to the run_queue state itself. This was done to simplify glock_work_func.
Eventually we want to reduce it to a single call, but this is just a
first step.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>