gfs2: Call state machine directly in gfs2_dispose_glock_lru
Now we start to reduce the use of the glock work queue.
In this patch, we replace the queueing of glock work in function
gfs2_dispose_glock_lru with a direct call to the state machine.
This requires us to release the lru_lock beforehand or we will
deadlock with states that may require lru_lock.
This will improve efficiency and reduce context switches.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>