gfs2: Allocate bd elements with each transaction
Before this patch, bd elements were allocated from slab memory as-needed
when each buffer is added to a transaction. Since it holds the log lock
and kmem_cache_alloc can block, it temporarily releases the locks and
reacquires them after the allocation. This is somewhat slow.

This patch makes each new transaction pre-allocate an array of bufdata
elements and puts them on a new free list, tr_free. As each buffer_head
is added to the transaction, we first check the tr_free list for free
elements, and if we have at least one, we dequeue and use it instead of
dropping the locks and doing a slow slab allocation.

This will save us the penalty of dropping and reacquiring the locks as
well as the slab allocation potentially blocking.

Each transaction is created with an estimate of how many blocks will be
required, we use this for the pre-allocation. Sometimes we don't use all
the journal log blocks we reserved during gfs2_trans_begin, so all
unused bd elements on the tr_free list are freed when the transaction is
destroyed.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>