gfs2: Simplify freeze by eliminating lfcc type
Before this patch, function gfs2_lock_fs_check_clean, which is a helper
for gfs2_freeze, had a special data structure called lfcc, which is
an array of glock holders, one for each journal to be locked.
This patch simplifies the code by including a glock holder with
each journal descriptor. That eliminates the need for the structure
and code necessary to allocate, manage and deallocate them.
This doesn't gain us much, but it helps us simplify the freeze/thaw
code with future patches.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>