gfs2: split do_demote into sync_inval and demote
Before this patch, the new state_demote set the state machine state
after it had unlocked the gl_lockref spin_lock. That's wrong because it
creates a race in which another process could change the state from what
we intended, which is FINISH_XMOTE.

This patch splits the do_demote state into two states: (1) one that
handles the page cache sync and, optionally, invalidation, and (2)
another that calls into the lock manager to change the lock state.
A new GLF_LM_LOCK bit is added to prevent multiple processing from
calling into lm_lock due to the split of the states.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>