gfs2: change refs to glock "state" to "mode"
This patch prepares for a transition to a finite state automaton to
manage glock work. It renames all glock and holder "state" variables
to "mode" variables. Since dlm uses "mode" it makes more sense to me
to have a glock held in "exclusive mode" rather than "exclusive state"
and so forth.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>