gfs2: Eliminate gl_target
There are too many glock mode fields, and it's confusing:
gl_mode, gl_demote_mode, gl_req, and gl_target. In an effort to
make it easier to understand, this patch eliminates the gl_target
field in favor of a new function, target_mode, that determines it
from the state of the glock and its holders.
Now we keep separated the lock mode we requested of DLM and the
reply we received from it.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>