gfs2: finish_xmote reform
Before this patch, finish_xmote was sometimes called to kind of pretend
we received a response from the lock manager in some cases, such as when
we use lock_nolock. This patch reroutes the code path so lock_nolock
calls into gfs2_glock_complete. Function mode_change has been eliminated
in favor of inlining it in gfs2_glock_complete. It also renames the
REPLY_PENDING bit to FINISH_XMOTE. This disentangles some of the complex
glock logic and makes it easier to do future improvements.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>