gfs2: Remove remote/queue work in handle_callback
Because of the previous patch, we now have only one caller to function
handle_callback. It always specified remote, so we no longer need to
pass that in as a parameter. Also, the only caller also queued glock
work. This patch moves that queuing into handle_callback to make the
code more understandable.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>