gfs2: Simplify state machine into blocking and non
The only time the state machine is now called is for a blocking
or non-blocking run_queue. This patch simplifies the state machine
so the initial state is always the run_queue state, with a flag to
indicate whether it may or may not block. That also means we eliminate
the "starting" state, which means it will be easier to always trace
back what happened when the state machine was last run.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>