gfs2: Control metapath buffer_head ownership
When cloning a metapath, we currently get references to all the buffer
heads it references.  Those references are put when the clone is
released.  Future patches will rely more heavily on cloned metapaths and
it will be important that cloning metapaths is efficient.  To achieve
that, use a bitmap in the metapath to indicate for which buffer heads
the metapath owns references.  A cloned metapath starts out not owning
any of the references.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>