gfs2: Improve gfs2_iomap_write_alloc transaction size estimate
The gfs2_write_calc_reserv() heuristics assume that writes won't span
more than one leaf indirect block.  On the way of removing this
limitation, stop using this function in gfs2_iomap_begin_write().
Instead, compute the actal number of data blocks required, and a bound
on the number of indirect blocks required for larger writes.

When unstuffing or writing into a hole, we know that we'll always have
to write the statfs inode, and we'll have to update quotas if they are
enabled.

Also, when unstuffing without writing into the first block, we will end
up writing an extra block.  Variable data_blocks accounts for that.
This is also the number of blocks we may need to revoke, unless it's a
jdata inode.  Pass that number to gfs2_trans_begin().

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>