gfs2: Convert __gfs2_fallocate to iomap
Convert __gfs2_fallocate() to iomap.  All the mapping functionality is
already there for iomap writes; the only difference is that fallocate
doesn't go through the page cache for zeroing out blocks.

This eliminates the last user of gfs2_iomap_alloc() that allocates
multiple blocks.  The two remaining users of gfs2_iomap_alloc() are in
the xattr and quota code, and the only allocate single blocks.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>