gfs2: Redefine RES_RG_BIT
The symbolic constant RES_RG_BIT is currently used to indicate the
number of blocks required for setting a bit in a resource group,
including updating the bitmap and the resource group header
(RES_RG_HDR).  Split the two so that RES_RG_BIT can be used to indicate
the number of bits when allocating multiple bits in the same resource
group.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>