gfs2: Get rid of repeated call to __gfs2_iomap_get
Right now, gfs2_iomap_begin() calls __gfs2_iomap_get(), and for writes,
it calls gfs2_iomap_begin_write().  When the inode needs to be
unstuffed, gfs2_iomap_begin_write() does that and then it calls
__gfs2_iomap_get() again to look up the mapping the unstuffing has
resulted in.

That works, but it isn't hard to figure out the mapping after unstuffing
without calling __gfs2_iomap_get() a second time, so do that instead.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>