gfs2: Make gfs2_glock_cb softirq safe
Make gfs2_glock_cb() softirq safe by splitting request_demote() into a
first part which can be called in softirq context and a second part
which needs to be run in process context.  Carry out the first part in
gfs2_glock_cb(), and the second part in gfs2_glock_work().

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>