gfs2: Rework asynchronous glock disposal mechanism
So far, the glock reference count has been tied to the number of times
glock work was run: each run of glock_work_func() consumed one
reference.  This was done so that glock_work_func() could be used to
drop the final reference.  This scheme has been working fine so far, but
it means that we need to take the glock spin lock for queuing glock
work.

In the future, we want to be able to queue glock work in contexts in
which we cannot take the glock spin lock.  To allow that, introduce a
new GLF_DROP_REF flag to indicate that glock_work_func() should consume
a glock reference and change glock_work_func() not to consume a
reference if that flag isn't set.  Setting GLF_DROP_REF still requires
taking the glock spin lock, but we can now queue additional glock work
without it.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>